chore: refactor and minor fixes on suitable position calculation

+++ positions-calculator-service/src/main/java/br/com/dgc/fmtools/positions_calculator_service/domain/model/position/SuitablePosition.java
@@ -2,6 +2,6 @@ package br.com.dgc.fmtools.positions_calculator_service.domain.model.position;
 
 public record SuitablePosition(
     String name,
-    Integer percentage,
+    Double percentage,
     Integer quantityKeyAttributes,
     Integer quantityPreferableAttributes) {}
